+++ src/calendar.js
@@ -4,7 +4,7 @@
  * A javascript drag & drop event calendar, that is fully responsive and compatible with all modern browsers.
  * 
  * @file        calendar.js
- * @version     v2.9.1
+ * @version     v2.9.2
  * @author      Bunoon
  * @license     GNU AGPLv3
  * @copyright   Bunoon 2023
@@ -1035,7 +1035,7 @@ function calendarJs( elementOrId, options, searchOptions ) {
         _initialized = true;
     
         if ( !_initialized_FirstTime ) {
-            triggerOptionsEventWithData( "onRender", _parameter_ElementID );
+            fireCustomTrigger( "onRender", _parameter_ElementID );
             _initialized_FirstTime = true;
         }
     }
@@ -1246,17 +1246,17 @@ function calendarJs( elementOrId, options, searchOptions ) {
 
     function saveSideMenuSelections() {
         if ( _element_SideMenu_Changed ) {
-            var triggerOptionsEvent = false,
+            var fireOptionsCustomTrigger = false,
                 itemWasChanged = false;
 
             if ( _element_SideMenu_Content_Section_Groups !== null ) {
                 var visibleGroups = getSideMenuCheckedCheckBoxNames( _element_SideMenu_Content_Section_Groups );
 
                 if ( !areArraysTheSame( _options_Configuration.visibleGroups, visibleGroups ) ) {
-                    _options_Configuration.visibleGroups = visibleGroups;
+                    _options_Configuration.visibleGroups = [].slice.call( visibleGroups );
                     itemWasChanged = true;
 
-                    triggerOptionsEventWithData( "onVisibleGroupsChanged", _options_Configuration.visibleGroups );
+                    fireCustomTrigger( "onVisibleGroupsChanged", _options_Configuration.visibleGroups );
                 }
             }
             
@@ -1264,10 +1264,10 @@ function calendarJs( elementOrId, options, searchOptions ) {
                 var visibleEventTypes = getSideMenuCheckedCheckBoxNames( _element_SideMenu_Content_Section_EventTypes, true );
 
                 if ( !areArraysTheSame( _options_Configuration.visibleEventTypes, visibleEventTypes ) ) {
-                    _options_Configuration.visibleEventTypes = visibleEventTypes;
+                    _options_Configuration.visibleEventTypes = [].slice.call( visibleEventTypes );
                     itemWasChanged = true;
 
-                    triggerOptionsEventWithData( "onVisibleEventTypesChanged", _options_Configuration.visibleEventTypes );
+                    fireCustomTrigger( "onVisibleEventTypesChanged", _options_Configuration.visibleEventTypes );
                 }
             }
     
@@ -1275,10 +1275,10 @@ function calendarJs( elementOrId, options, searchOptions ) {
                 var visibleDays = getSideMenuCheckedCheckBoxNames( _element_SideMenu_Content_Section_Days, true );
 
                 if ( visibleDays.length >= 1 && !areArraysTheSame( _options.visibleDays, visibleDays ) ) {
-                    _options.visibleDays = visibleDays;
+                    _options.visibleDays = [].slice.call( visibleDays );
                     _element_Calendar_PreviousDaysVisibleBeforeSingleDayView = [];
                     
-                    triggerOptionsEvent = true;
+                    fireOptionsCustomTrigger = true;
                     itemWasChanged = true;
                 }
             }
@@ -1287,9 +1287,9 @@ function calendarJs( elementOrId, options, searchOptions ) {
                 var workingDays = getSideMenuCheckedCheckBoxNames( _element_SideMenu_Content_Section_WorkingDays, true );
 
                 if ( !areArraysTheSame( _options.workingDays, workingDays ) ) {
-                    _options.workingDays = workingDays;
+                    _options.workingDays = [].slice.call( workingDays );
 
-                    triggerOptionsEvent = true;
+                    fireOptionsCustomTrigger = true;
                     itemWasChanged = true;
                 }
             }
@@ -1300,14 +1300,14 @@ function calendarJs( elementOrId, options, searchOptions ) {
                 if ( !areArraysTheSame( _options.weekendDays, weekendDays ) ) {
                     _options.weekendDays = weekendDays;
 
-                    triggerOptionsEvent = true;
+                    fireOptionsCustomTrigger = true;
                     itemWasChanged = true;
                 }
             }
 
             if ( itemWasChanged ) {
-                if ( triggerOptionsEvent ) {
-                    triggerOptionsEventWithData( "onOptionsUpdated", _options );
+                if ( fireOptionsCustomTrigger ) {
+                    fireCustomTrigger( "onOptionsUpdated", _options );
                 }
         
                 _initialized = false;
@@ -1676,12 +1676,13 @@ function calendarJs( elementOrId, options, searchOptions ) {
         events.appendChild( event );
 
         setEventClassesAndColors( event, eventDetails, getToTimeWithPassedDate( eventDetails, _calendar_CurrentDate ) );
+        setEventCustomTriggers( event, eventDetails );
 
         event.id = _element_ID_Event_Widget_Day + eventDetails.id;
         event.setAttribute( "event-type", getNumber( eventDetails.type ) );
         event.setAttribute( "event-id", eventDetails.id );
 
-        if ( !triggerOptionsEventWithMultipleData( "onWidgetEventRender", event, eventDetails ) ) {
+        if ( !fireCustomTrigger( "onWidgetEventRender", event, eventDetails ) ) {
             var title = createElement( "div", "title" ),
                 repeatEvery = getNumber( eventDetails.repeatEvery );
 
@@ -1733,26 +1734,6 @@ function calendarJs( elementOrId, options, searchOptions ) {
                 var description = createElement( "div", "description" );
                 setNodeText( description, eventDetails.description );
                 event.appendChild( description );
-            }
-        }
-
-        if ( isOptionEventSet( "onEventClick" ) ) {
-            event.addEventListener( "click", function() {
-                triggerOptionsEventWithData( "onEventClick", eventDetails );
-            } );
-        }
-
-        if ( _options.manualEditingEnabled ) {
-            event.ondblclick = function( e ) {
-                cancelBubble( e );
-                showEventEditingDialog( eventDetails );
-            };
-        } else {
-
-            if ( isOptionEventSet( "onEventDoubleClick" ) ) {
-                event.ondblclick = function() {
-                    triggerOptionsEventWithData( "onEventDoubleClick", eventDetails );
-                };
             }
         }
 
@@ -1845,14 +1826,14 @@ function calendarJs( elementOrId, options, searchOptions ) {
             _element_Calendar.className = "calendar calendar-shown";
 
             build( new Date( _calendar_CurrentDate_ForDatePicker ), !_initialized );
-            triggerOptionsEventWithData( "onDatePickerOpened", _parameter_ElementID );
+            fireCustomTrigger( "onDatePickerOpened", _parameter_ElementID );
             updateDatePickerPosition();
         } else {
 
             _element_Calendar.className = "calendar calendar-hidden";
             
             hideAllDropDowns();
-            triggerOptionsEventWithData( "onDatePickerClosed", _parameter_ElementID );
+            fireCustomTrigger( "onDatePickerClosed", _parameter_ElementID );
         }
 
         _element_Mode_DatePicker_Visible = !_element_Mode_DatePicker_Visible;
@@ -1892,7 +1873,7 @@ function calendarJs( elementOrId, options, searchOptions ) {
             _element_Mode_DatePicker_Visible = false;
 
             hideAllDropDowns();
-            triggerOptionsEventWithData( "onDatePickerClosed", _parameter_ElementID );
+            fireCustomTrigger( "onDatePickerClosed", _parameter_ElementID );
         }
     }
 
@@ -1905,7 +1886,7 @@ function calendarJs( elementOrId, options, searchOptions ) {
             
             hideDatePickerMode();
             updateDatePickerInputValueDisplay( date );
-            triggerOptionsEventWithData( "onDatePickerDateChanged", newDate );
+            fireCustomTrigger( "onDatePickerDateChanged", newDate );
 
             _calendar_CurrentDate_ForDatePicker = newDate;
             
@@ -2529,12 +2510,13 @@ function calendarJs( elementOrId, options, searchOptions ) {
     
             setEventClassesForActions( event, eventDetails );
             setEventClassesAndColors( event, eventDetails, getToTimeWithPassedDate( eventDetails, displayDate ) );
+            setEventCustomTriggers( event, eventDetails, editEventDate );
 
             if ( doDatesMatch( eventDetails.from, displayDate ) ) {
                 event.id = _element_ID_Event_FullDay + eventDetails.id;
             }
 
-            if ( !triggerOptionsEventWithMultipleData( "onFullDayEventRender", event, eventDetails ) ) {
+            if ( !fireCustomTrigger( "onFullDayEventRender", event, eventDetails ) ) {
                 var title = createElement( "div", "title" ),
                     repeatEvery = getNumber( eventDetails.repeatEvery );
 
@@ -2594,26 +2576,6 @@ function calendarJs( elementOrId, options, searchOptions ) {
             event.addEventListener( "click", function( e ) {
                 storeMultiSelectEvent( e, eventDetails );
             } );
-
-            if ( isOptionEventSet( "onEventClick" ) ) {
-                event.addEventListener( "click", function() {
-                    triggerOptionsEventWithData( "onEventClick", eventDetails );
-                } );
-            }
-    
-            if ( _options.manualEditingEnabled ) {
-                event.ondblclick = function( e ) {
-                    cancelBubble( e );
-                    showEventEditingDialog( eventDetails, null, null, editEventDate );
-                };
-            } else {
-
-                if ( isOptionEventSet( "onEventDoubleClick" ) ) {
-                    event.ondblclick = function() {
-                        triggerOptionsEventWithData( "onEventDoubleClick", eventDetails );
-                    };
-                }
-            }
             
             if ( !eventDetails.isAllDay ) {
                 scrollTop = setEventPositionAndGetScrollTop( _element_View_FullDay_Contents, _element_View_FullDay_Contents_Hours, displayDate, event, eventDetails );
@@ -2642,7 +2604,7 @@ function calendarJs( elementOrId, options, searchOptions ) {
     }
 
     function buildFullDayViewTitle() {
-        if ( !triggerOptionsEventWithData( "onFullDayTitleRender", _element_View_FullDay_DateSelected ) ) {
+        if ( !fireCustomTrigger( "onFullDayTitleRender", _element_View_FullDay_DateSelected ) ) {
             buildDateTimeDisplay( _element_View_FullDay_TitleBar, _element_View_FullDay_DateSelected, false, true, true );
         }
     }
@@ -3049,7 +3011,7 @@ function calendarJs( elementOrId, options, searchOptions ) {
     function buildFullWeekViewTitle( weekStartDate, weekEndDate ) {
         _element_View_FullWeek_TitleBar.innerHTML = _string.empty;
 
-        if ( !triggerOptionsEventWithMultipleData( "onFullWeekTitleRender", weekStartDate, weekEndDate ) ) {
+        if ( !fireCustomTrigger( "onFullWeekTitleRender", weekStartDate, weekEndDate ) ) {
             if ( _options.showWeekNumbersInTitles ) {
                 createSpanElement( _element_View_FullWeek_TitleBar, _options.weekText + _string.space + getWeekNumber( weekStartDate ) + ": " );
             }
@@ -3157,12 +3119,13 @@ function calendarJs( elementOrId, options, searchOptions ) {
             
             setEventClassesForActions( event, eventDetails );
             setEventClassesAndColors( event, eventDetails, getToTimeWithPassedDate( eventDetails, actualDisplayDate ) );
+            setEventCustomTriggers( event, eventDetails, editEventDate );
 
             if ( doDatesMatch( eventDetails.from, actualDisplayDate ) ) {
                 event.id = _element_ID_Event_WeekDay + eventDetails.id;
             }
 
-            if ( !triggerOptionsEventWithMultipleData( "onFullWeekEventRender", event, eventDetails ) ) {
+            if ( !fireCustomTrigger( "onFullWeekEventRender", event, eventDetails ) ) {
                 var title = createElement( "div", "title" ),
                     repeatEvery = getNumber( eventDetails.repeatEvery );
 
@@ -3222,26 +3185,6 @@ function calendarJs( elementOrId, options, searchOptions ) {
             event.addEventListener( "click", function( e ) {
                 storeMultiSelectEvent( e, eventDetails );
             } );
-
-            if ( isOptionEventSet( "onEventClick" ) ) {
-                event.addEventListener( "click", function() {
-                    triggerOptionsEventWithData( "onEventClick", eventDetails );
-                } );
-            }
-    
-            if ( _options.manualEditingEnabled ) {
-                event.ondblclick = function( e ) {
-                    cancelBubble( e );
-                    showEventEditingDialog( eventDetails, null, null, editEventDate );
-                };
-            } else {
-
-                if ( isOptionEventSet( "onEventDoubleClick" ) ) {
-                    event.ondblclick = function() {
-                        triggerOptionsEventWithData( "onEventDoubleClick", eventDetails );
-                    };
-                }
-            }
 
             if ( !eventDetails.isAllDay ) {
                 setEventPositionAndGetScrollTop( column, column, actualDisplayDate, event, eventDetails, true );
@@ -3651,7 +3594,7 @@ function calendarJs( elementOrId, options, searchOptions ) {
     }
 
     function buildFullMonthViewPinUpImageText() {
-        if ( !_element_Mode_DatePicker_Enabled && _element_View_FullMonth_PinUp !== null && !triggerOptionsEventWithMultipleData( "onFullMonthPinUpRender", _element_View_FullMonth_PinUp, _calendar_CurrentDate ) ) {
+        if ( !_element_Mode_DatePicker_Enabled && _element_View_FullMonth_PinUp !== null && !fireCustomTrigger( "onFullMonthPinUpRender", _element_View_FullMonth_PinUp, _calendar_CurrentDate ) ) {
             var currentDate = new Date();
 
             if ( _element_View_FullMonth_PinUp_CurrentDate === null || !doDatesMatch( _element_View_FullMonth_PinUp_CurrentDate, currentDate ) ) {
@@ -4312,7 +4255,7 @@ function calendarJs( elementOrId, options, searchOptions ) {
                         event.setAttribute( "event-id", eventDetails.id );
 
                         if ( !_options.useOnlyDotEventsForMainDisplay ) {
-                            if ( !triggerOptionsEventWithMultipleData( "onFullMonthEventRender", event, eventDetails ) ) {
+                            if ( !fireCustomTrigger( "onFullMonthEventRender", event, eventDetails ) ) {
                                 var eventTitle = eventDetails.title,
                                     repeatEvery = getNumber( eventDetails.repeatEvery );
             
@@ -4340,7 +4283,8 @@ function calendarJs( elementOrId, options, searchOptions ) {
                         makeEventDraggable( event, eventDetails, dayDate, elementDay );
                         setEventClassesForActions( event, eventDetails );
                         setEventClassesAndColors( event, eventDetails, getToTimeWithPassedDate( eventDetails, dayDate ), _options.applyCssToEventsNotInCurrentMonth );
-    
+                        setEventCustomTriggers( event, eventDetails, editEventDate );
+
                         if ( doDatesMatch( eventDetails.from, dayDate ) ) {
                             event.id = _element_ID_Event_Day + eventDetails.id;
                         }
@@ -4360,26 +4304,6 @@ function calendarJs( elementOrId, options, searchOptions ) {
                         event.addEventListener( "click", function( e ) {
                             storeMultiSelectEvent( e, eventDetails );
                         } );
-    
-                        if ( isOptionEventSet( "onEventClick" ) ) {
-                            event.addEventListener( "click", function() {
-                                triggerOptionsEventWithData( "onEventClick", eventDetails );
-                            } );
-                        }
-            
-                        if ( _options.manualEditingEnabled ) {
-                            event.ondblclick = function( e ) {
-                                cancelBubble( e );
-                                showEventEditingDialog( eventDetails, null, null, editEventDate );
-                            };
-                        } else {
-    
-                            if ( isOptionEventSet( "onEventDoubleClick" ) ) {
-                                event.ondblclick = function() {
-                                    triggerOptionsEventWithData( "onEventDoubleClick", eventDetails );
-                                };
-                            }
-                        }
         
                     } else {
                         buildFullMonthViewDayEventPlusText( elementDay, dayDate );
@@ -4478,7 +4402,7 @@ function calendarJs( elementOrId, options, searchOptions ) {
         if ( _calendar_IsBusy_LastState !== _calendar_IsBusy ) {
             _calendar_IsBusy_LastState = _calendar_IsBusy;
 
-            triggerOptionsEventWithData( "onBusyStateChange", _calendar_IsBusy );
+            fireCustomTrigger( "onBusyStateChange", _calendar_IsBusy );
         }
     }
 
@@ -4909,12 +4833,13 @@ function calendarJs( elementOrId, options, searchOptions ) {
             makeEventDraggable( event, eventDetails, eventDetails.from, container );
             setEventClassesForActions( event, eventDetails );
             setEventClassesAndColors( event, eventDetails );
+            setEventCustomTriggers( event, eventDetails );
 
             event.id = _element_ID_Event_Month + eventDetails.id;
             event.setAttribute( "event-type", getNumber( eventDetails.type ) );
             event.setAttribute( "event-id", eventDetails.id );
 
-            if ( !triggerOptionsEventWithMultipleData( "onAllEventsEventRender", event, eventDetails ) ) {
+            if ( !fireCustomTrigger( "onAllEventsEventRender", event, eventDetails ) ) {
                 var title = createElement( "div", "title" ),
                     repeatEvery = getNumber( eventDetails.repeatEvery );
     
@@ -4972,26 +4897,6 @@ function calendarJs( elementOrId, options, searchOptions ) {
             event.addEventListener( "click", function( e ) {
                 storeMultiSelectEvent( e, eventDetails );
             } );
-
-            if ( isOptionEventSet( "onEventClick" ) ) {
-                event.addEventListener( "click", function() {
-                    triggerOptionsEventWithData( "onEventClick", eventDetails );
-                } );
-            }
-    
-            if ( _options.manualEditingEnabled ) {
-                event.ondblclick = function( e ) {
-                    cancelBubble( e );
-                    showEventEditingDialog( eventDetails );
-                };
-            } else {
-
-                if ( isOptionEventSet( "onEventDoubleClick" ) ) {
-                    event.ondblclick = function() {
-                        triggerOptionsEventWithData( "onEventDoubleClick", eventDetails );
-                    };
-                }
-            }
 
             _element_View_AllEvents_EventsShown.push( eventDetails );
         }
@@ -5352,14 +5257,15 @@ function calendarJs( elementOrId, options, searchOptions ) {
 
     function buildTimelineViewEvent( timelineRowItems, eventDetails ) {
         var event = createElement( "div", "event" ),
-            formattedDate = toStorageFormattedDate( _element_View_Timeline_DateSelected );
+            formattedDate = toStorageFormattedDate( _element_View_Timeline_DateSelected ),
+            editEventDate = new Date( _element_View_Timeline_DateSelected );
         
         event.id = _element_ID_Event_Timeline_Day + eventDetails.id;
         event.setAttribute( "event-type", getNumber( eventDetails.type ) );
         event.setAttribute( "event-id", eventDetails.id );
         timelineRowItems.appendChild( event );
 
-        if ( !triggerOptionsEventWithMultipleData( "onTimelineEventRender", event, eventDetails ) ) {
+        if ( !fireCustomTrigger( "onTimelineEventRender", event, eventDetails ) ) {
             var repeatEvery = getNumber( eventDetails.repeatEvery );
 
             if ( repeatEvery > _enum_RepeatType.never ) {
@@ -5387,26 +5293,7 @@ function calendarJs( elementOrId, options, searchOptions ) {
             }
         };
 
-        if ( isOptionEventSet( "onEventClick" ) ) {
-            event.addEventListener( "click", function() {
-                triggerOptionsEventWithData( "onEventClick", eventDetails );
-            } );
-        }
-
-        if ( _options.manualEditingEnabled ) {
-            event.ondblclick = function( e ) {
-                cancelBubble( e );
-                showEventEditingDialog( eventDetails, null, null, _element_View_Timeline_DateSelected );
-            };
-        } else {
-
-            if ( isOptionEventSet( "onEventDoubleClick" ) ) {
-                event.ondblclick = function() {
-                    triggerOptionsEventWithData( "onEventDoubleClick", eventDetails );
-                };
-            }
-        }
-
+        setEventCustomTriggers( event, eventDetails, editEventDate );
         setEventClassesForActions( event, eventDetails );
         setEventClassesAndColors( event, eventDetails, getToTimeWithPassedDate( eventDetails, _element_View_Timeline_DateSelected ) );
         
@@ -5426,7 +5313,7 @@ function calendarJs( elementOrId, options, searchOptions ) {
     }
 
     function buildTimelineViewTitle() {
-        if ( !triggerOptionsEventWithData( "onTimelineTitleRender", _element_View_FullDay_DateSelected ) ) {
+        if ( !fireCustomTrigger( "onTimelineTitleRender", _element_View_FullDay_DateSelected ) ) {
             buildDateTimeDisplay( _element_View_Timeline_TitleBar, _element_View_Timeline_DateSelected, false, true, true );
         }
     }
@@ -6001,7 +5888,7 @@ function calendarJs( elementOrId, options, searchOptions ) {
                 _options.visibleDays.splice( _options.visibleDays.indexOf( _element_ContextMenu_HeaderDay_SelectedDay ), 1 );
                 _initialized = false;
     
-                triggerOptionsEventWithData( "onOptionsUpdated", _options );
+                fireCustomTrigger( "onOptionsUpdated", _options );
                 build( _calendar_CurrentDate, true, true );
             }, true );
     
@@ -6012,7 +5899,7 @@ function calendarJs( elementOrId, options, searchOptions ) {
                     _options.visibleDays = [].slice.call( _options.workingDays );
                     _initialized = false;
         
-                    triggerOptionsEventWithData( "onOptionsUpdated", _options );
+                    fireCustomTrigger( "onOptionsUpdated", _options );
                     build( _calendar_CurrentDate, true, true );
                 }
             } );
@@ -6353,7 +6240,7 @@ function calendarJs( elementOrId, options, searchOptions ) {
 
     function showEventEditingDialog( eventDetails, overrideTodayDate, overrideTimeValues, originDayDate ) {
         if ( isFunction( _options.onBeforeEventAddEdit ) ) {
-            triggerOptionsEventWithData( "onBeforeEventAddEdit", eventDetails );
+            fireCustomTrigger( "onBeforeEventAddEdit", eventDetails );
         } else {
             
             addNode( _parameter_Document.body, _element_Calendar_DisabledBackground );
@@ -6655,7 +6542,7 @@ function calendarJs( elementOrId, options, searchOptions ) {
                         } else {
                             _this.addEvent( newEvent, false );
         
-                            showNotificationPopUp( _options.eventAddedText.replace( "{0}", _element_Dialog_EventEditor_EventDetails.title ) );
+                            showNotificationPopUp( _options.eventAddedText.replace( "{0}", newEvent.title ) );
                         }
 
                         if ( _options.isWidget ) {
@@ -7657,7 +7544,7 @@ function calendarJs( elementOrId, options, searchOptions ) {
                     _options_Search.top = _element_Dialog_Search.offsetTop;
                 }
     
-                triggerOptionsEventWithData( "onSearchOptionsUpdated", _options_Search );
+                fireCustomTrigger( "onSearchOptionsUpdated", _options_Search );
             }
         }, 2000, false );
     }
@@ -7882,7 +7769,7 @@ function calendarJs( elementOrId, options, searchOptions ) {
 
         _initialized = false;
 
-        triggerOptionsEventWithData( "onOptionsUpdated", _options );
+        fireCustomTrigger( "onOptionsUpdated", _options );
         checkForBrowserNotificationsPermission();
         hideConfigurationDialog();
         build( _calendar_CurrentDate, true, true );
@@ -7997,7 +7884,7 @@ function calendarJs( elementOrId, options, searchOptions ) {
                         _element_Tooltip_Title.innerHTML = _string.empty;
                         _element_Tooltip_TotalTime.innerHTML = _string.empty;
 
-                        if ( !triggerOptionsEventWithMultipleData( "onToolTipEventRender", _element_Tooltip, eventDetails ) ) {
+                        if ( !fireCustomTrigger( "onToolTipEventRender", _element_Tooltip, eventDetails ) ) {
                             _element_Tooltip.onmousemove = cancelBubble;
                             _element_Tooltip.appendChild( _element_Tooltip_TitleButtons );
                             _element_Tooltip.appendChild( _element_Tooltip_Title );
@@ -8230,7 +8117,7 @@ function calendarJs( elementOrId, options, searchOptions ) {
             if ( updateDisplay ) {
                 _initialized = false;
     
-                triggerOptionsEventWithData( "onOptionsUpdated", _options );
+                fireCustomTrigger( "onOptionsUpdated", _options );
                 build( _calendar_CurrentDate, true, true );
             }
         }
@@ -8642,7 +8529,7 @@ function calendarJs( elementOrId, options, searchOptions ) {
             _element_View_Event_Dragged_EventDetails.to = addMinutesToDate( _element_View_Event_Dragged_EventDetails.to, differenceMinutes );
             
             storeEventsInLocalStorage();
-            triggerOptionsEventWithData( "onEventUpdated", _element_View_Event_Dragged_EventDetails );
+            fireCustomTrigger( "onEventUpdated", _element_View_Event_Dragged_EventDetails );
             showNotificationPopUp( _options.eventUpdatedText.replace( "{0}", _element_View_Event_Dragged_EventDetails.title ) );
             refreshOpenedViews();
 
@@ -8686,7 +8573,7 @@ function calendarJs( elementOrId, options, searchOptions ) {
                         eventSizeDetails.eventDetails.to = addMinutesToDate( eventSizeDetails.eventDetails.to, differenceMinutes );
                         eventsResized = true;
 
-                        triggerOptionsEventWithData( "onEventUpdated", eventSizeDetails.eventDetails );
+                        fireCustomTrigger( "onEventUpdated", eventSizeDetails.eventDetails );
                         showNotificationPopUp( _options.eventUpdatedText.replace( "{0}", eventSizeDetails.eventDetails.title ) );
                     }
                 }
@@ -8717,7 +8604,7 @@ function calendarJs( elementOrId, options, searchOptions ) {
     function turnOnFullScreenMode() {
         if ( !_element_Calendar_FullScreenModeOn && _options.fullScreenModeEnabled ) {
             forceTurnOnFullScreenMode();
-            triggerOptionsEventWithData( "onFullScreenModeChanged", true );
+            fireCustomTrigger( "onFullScreenModeChanged", true );
         }
     }
 
@@ -8729,7 +8616,7 @@ function calendarJs( elementOrId, options, searchOptions ) {
 
             updateFullScreenModeExpandButtons( "ib-arrow-expand-left-right", _options.enableFullScreenTooltipText );
             refreshOpenedViews();
-            triggerOptionsEventWithData( "onFullScreenModeChanged", false );
+            fireCustomTrigger( "onFullScreenModeChanged", false );
         }
     }
 
@@ -9412,7 +9299,7 @@ function calendarJs( elementOrId, options, searchOptions ) {
             event.setAttribute( "draggable", true );
             
             event.ondragstart = function( e ) {
-                triggerOptionsEventWithData( "onEventDragStart", eventDetails );
+                fireCustomTrigger( "onEventDragStart", eventDetails );
 
                 e.dataTransfer.setData( "event_details", JSON.stringify( eventDetails ) );
 
@@ -9435,7 +9322,7 @@ function calendarJs( elementOrId, options, searchOptions ) {
             };
 
             event.ondragend = function() {
-                triggerOptionsEventWithData( "onEventDragStop", _events_Dragged );
+                fireCustomTrigger( "onEventDragStop", _events_Dragged );
 
                 _events_Dragged_DateFrom = null;
                 _events_Dragged = null;
@@ -9515,7 +9402,7 @@ function calendarJs( elementOrId, options, searchOptions ) {
         var dropDate = new Date( year, month, day );
 
         if ( _events_Dragged !== null && !doDatesMatch( _events_Dragged_DateFrom, dropDate ) ) {
-            triggerOptionsEventWithMultipleData( "onEventDragDrop", _events_Dragged, dropDate );
+            fireCustomTrigger( "onEventDragDrop", _events_Dragged, dropDate );
 
             if ( !isDefined( day ) ) {
                 var totalDaysInMonth = getTotalDaysInMonth( year, month );
@@ -9877,7 +9764,7 @@ function calendarJs( elementOrId, options, searchOptions ) {
                 openEventUrl( url );
             }
 
-            triggerOptionsEventWithData( "onNotificationClicked", eventDetails );
+            fireCustomTrigger( "onNotificationClicked", eventDetails );
         };
     }
 
@@ -9902,7 +9789,7 @@ function calendarJs( elementOrId, options, searchOptions ) {
             }
 
             if ( isDefined( eventDetails ) ) {
-                triggerOptionsEventWithData( "onNotification", eventDetails );
+                fireCustomTrigger( "onNotification", eventDetails );
             }
         }
     }
@@ -9910,7 +9797,7 @@ function calendarJs( elementOrId, options, searchOptions ) {
     function openEventUrl( url ) {
         _parameter_Window.open( url, _options.urlWindowTarget );
 
-        triggerOptionsEvent( "onEventUrlClicked", url );
+        fireCustomTrigger( "onEventUrlClicked", url );
     }
 
 
@@ -10001,7 +9888,7 @@ function calendarJs( elementOrId, options, searchOptions ) {
                 _this.addEvent( newEvent, false, true );
 
             } else {
-                triggerOptionsEventWithData( "onEventUpdated", newEvent );
+                fireCustomTrigger( "onEventUpdated", newEvent );
             }
         }
 
@@ -10109,7 +9996,7 @@ function calendarJs( elementOrId, options, searchOptions ) {
     }
 
     function loadEventsToAddOrUpdateFromFetchTrigger() {
-        var events = triggerOptionsEvent( "onEventsFetch" );
+        var events = fireCustomTrigger( "onEventsFetch" );
         if ( isDefinedArray( events ) ) {
             var eventsLength = events.length;
 
@@ -10151,7 +10038,7 @@ function calendarJs( elementOrId, options, searchOptions ) {
                 }
     
                 if ( triggerEvent ) {
-                    triggerOptionsEvent( "onRefresh" );
+                    fireCustomTrigger( "onRefresh" );
                 }
             }
         }
@@ -11153,7 +11040,7 @@ function calendarJs( elementOrId, options, searchOptions ) {
             buildFullMonthViewDayEvents();
             refreshOpenedViews();
             showNotificationPopUp( _options.eventsImportedText.replace( "{0}", eventsAddedOrUpdated.length ) );
-            triggerOptionsEventWithData( "onEventsImported", eventsAddedOrUpdated );
+            fireCustomTrigger( "onEventsImported", eventsAddedOrUpdated );
         }
     }
 
@@ -11216,7 +11103,7 @@ function calendarJs( elementOrId, options, searchOptions ) {
                 showNotificationPopUp( _options.eventsExportedToText.replace( "{0}", filename ) );
             }
 
-            triggerOptionsEventWithData( "onEventsExported", contentsEvents );
+            fireCustomTrigger( "onEventsExported", contentsEvents );
         }
     }
 
@@ -12030,38 +11917,50 @@ function calendarJs( elementOrId, options, searchOptions ) {
      * ----------------------------------------------------------------------------------------------------------------------------------------------------------------------------
      */
 
-    function isOptionEventSet( name ) {
+    function isCustomTriggerSet( name ) {
         return isDefinedFunction( _options[ name ] );
     }
 
-    function triggerOptionsEvent( name ) {
-        var result = null;
+    function fireCustomTrigger( name ) {
+        var result = null,
+            newArguments = [].slice.call( arguments, 1 );
 
-        if ( _options !== null && isOptionEventSet( name ) ) {
-            result = _options[ name ]();
+        if ( newArguments.length > 0 ) {
+            result = false;
+        }
+        
+        if ( _options !== null && isCustomTriggerSet( name ) ) {
+            result = _options[ name ].apply( null, newArguments );
         }
 
         return result;
     }
 
-    function triggerOptionsEventWithData( name, data ) {
-        var result = false;
-
-        if ( _options !== null && isOptionEventSet( name ) ) {
-            result = _options[ name ]( data );
+    function setEventCustomTriggers( event, eventDetails, editEventDate ) {
+        if ( isCustomTriggerSet( "onEventClick" ) ) {
+            event.addEventListener( "click", function() {
+                fireCustomTrigger( "onEventClick", eventDetails );
+            } );
         }
 
-        return result;
-    }
+        if ( _options.manualEditingEnabled ) {
+            event.ondblclick = function( e ) {
+                cancelBubble( e );
 
-    function triggerOptionsEventWithMultipleData( name, data1, data2 ) {
-        var result = false;
+                if ( isDefined( editEventDate ) ) {
+                    showEventEditingDialog( eventDetails, null, null, editEventDate );
+                } else {
+                    showEventEditingDialog( eventDetails );
+                }
+            };
+        } else {
 
-        if ( _options !== null && isOptionEventSet( name ) ) {
-            result = _options[ name ]( data1, data2 );
+            if ( isCustomTriggerSet( "onEventDoubleClick" ) ) {
+                event.ondblclick = function() {
+                    fireCustomTrigger( "onEventDoubleClick", eventDetails );
+                };
+            }
         }
-
-        return result;
     }
 
 
@@ -12184,7 +12083,7 @@ function calendarJs( elementOrId, options, searchOptions ) {
             _element_Calendar.innerHTML = _string.empty;
         }
 
-        triggerOptionsEvent( "onDestroy", _parameter_ElementID );
+        fireCustomTrigger( "onDestroy", _parameter_ElementID );
 
         return this;
     };
@@ -12308,7 +12207,7 @@ function calendarJs( elementOrId, options, searchOptions ) {
             if ( !doDatesMatch( _calendar_CurrentDate, newDate ) ) {
                 if ( newDate.getFullYear() >= _options.minimumYear && newDate.getFullYear() <= _options.maximumYear ) {
                     build( newDate );
-                    triggerOptionsEventWithData( "onSetDate", newDate );
+                    fireCustomTrigger( "onSetDate", newDate );
                 }
             }
         }
@@ -12352,7 +12251,7 @@ function calendarJs( elementOrId, options, searchOptions ) {
 
                     hideDatePickerMode();
                     updateDatePickerInputValueDisplay( newDate );
-                    triggerOptionsEventWithData( "onDatePickerDateChanged", newDate );
+                    fireCustomTrigger( "onDatePickerDateChanged", newDate );
     
                     _calendar_CurrentDate_ForDatePicker = newDate;
                 }
@@ -12413,7 +12312,7 @@ function calendarJs( elementOrId, options, searchOptions ) {
 
             if ( previousMonth.getFullYear() >= _options.minimumYear ) {
                 build( previousMonth );
-                triggerOptionsEventWithData( "onPreviousMonth", previousMonth );
+                fireCustomTrigger( "onPreviousMonth", previousMonth );
                 buildFullMonthViewPinUpImage();
             }
         }
@@ -12430,7 +12329,7 @@ function calendarJs( elementOrId, options, searchOptions ) {
 
             if ( nextMonth.getFullYear() <= _options.maximumYear ) {
                 build( nextMonth );
-                triggerOptionsEventWithData( "onNextMonth", nextMonth );
+                fireCustomTrigger( "onNextMonth", nextMonth );
                 buildFullMonthViewPinUpImage();
             }
         }
@@ -12443,7 +12342,7 @@ function calendarJs( elementOrId, options, searchOptions ) {
     
             if ( previousYear.getFullYear() >= _options.minimumYear ) {
                 build( previousYear );
-                triggerOptionsEventWithData( "onPreviousYear", previousYear );
+                fireCustomTrigger( "onPreviousYear", previousYear );
                 buildFullMonthViewPinUpImage();
             }
         }
@@ -12456,7 +12355,7 @@ function calendarJs( elementOrId, options, searchOptions ) {
 
             if ( nextYear.getFullYear() <= _options.maximumYear ) {
                 build( nextYear );
-                triggerOptionsEventWithData( "onNextYear", nextYear );
+                fireCustomTrigger( "onNextYear", nextYear );
                 buildFullMonthViewPinUpImage();
             }
         }
@@ -12468,7 +12367,7 @@ function calendarJs( elementOrId, options, searchOptions ) {
 
             if ( _calendar_CurrentDate.getMonth() !== today.getMonth() || _calendar_CurrentDate.getFullYear() !== today.getFullYear() ) {
                 build();
-                triggerOptionsEvent( "onToday" );
+                fireCustomTrigger( "onToday" );
                 buildFullMonthViewPinUpImage();
             }
         }
@@ -12503,7 +12402,7 @@ function calendarJs( elementOrId, options, searchOptions ) {
             this.addEvents( events, updateEvents, false );
     
             if ( triggerEvent ) {
-                triggerOptionsEventWithData( "onEventsSet", events );
+                fireCustomTrigger( "onEventsSet", events );
             }
         }
 
@@ -12537,7 +12436,7 @@ function calendarJs( elementOrId, options, searchOptions ) {
             }
     
             if ( triggerEvent ) {
-                triggerOptionsEventWithData( "onEventsSetFromJSON", json );
+                fireCustomTrigger( "onEventsSetFromJSON", json );
             }
         }
 
@@ -12573,7 +12472,7 @@ function calendarJs( elementOrId, options, searchOptions ) {
             storeEventsInLocalStorage();
     
             if ( triggerEvent ) {
-                triggerOptionsEventWithData( "onEventsAdded", events );
+                fireCustomTrigger( "onEventsAdded", events );
             }
     
             if ( updateEvents ) {
@@ -12613,7 +12512,7 @@ function calendarJs( elementOrId, options, searchOptions ) {
             }
     
             if ( triggerEvent ) {
-                triggerOptionsEventWithData( "onEventsAddedFromJSON", json );
+                fireCustomTrigger( "onEventsAddedFromJSON", json );
             }
         }
 
@@ -12691,8 +12590,12 @@ function calendarJs( elementOrId, options, searchOptions ) {
                         group = getString( event.group ),
                         url = getString( event.url );
 
-                    if ( isDefined( _options_Configuration.visibleGroups ) ) {
-                        _options_Configuration.visibleGroups.push( getGroupName( group ) );
+                    if ( isDefined( _options_Configuration.visibleGroups ) && !isCustomTriggerSet( "onEventsFetch" ) ) {
+                        var groupName = getGroupName( group );
+
+                        if ( _options_Configuration.visibleGroups.indexOf( groupName ) === -1 ) {
+                            _options_Configuration.visibleGroups.push( groupName );
+                        }
                     }
     
                     if ( !isDefined( event.id ) ) {
@@ -12733,7 +12636,7 @@ function calendarJs( elementOrId, options, searchOptions ) {
                     added = true;
     
                     if ( triggerEvent ) {
-                        triggerOptionsEventWithData( "onEventAdded", event );
+                        fireCustomTrigger( "onEventAdded", event );
                     }
             
                     if ( updateEvents ) {
@@ -12776,7 +12679,7 @@ function calendarJs( elementOrId, options, searchOptions ) {
             }
     
             if ( triggerEvent ) {
-                triggerOptionsEventWithData( "onEventsUpdated", events );
+                fireCustomTrigger( "onEventsUpdated", events );
             }
     
             if ( updateEvents ) {
@@ -12820,7 +12723,7 @@ function calendarJs( elementOrId, options, searchOptions ) {
                 storeEventsInLocalStorage();
     
                 if ( updated && triggerEvent ) {
-                    triggerOptionsEventWithData( "onEventUpdated", event );
+                    fireCustomTrigger( "onEventUpdated", event );
                 }
             }
         }
@@ -12860,7 +12763,7 @@ function calendarJs( elementOrId, options, searchOptions ) {
                     updated = true;
     
                     if ( triggerEvent ) {
-                        triggerOptionsEventWithData( "onEventUpdated", eventDetails );
+                        fireCustomTrigger( "onEventUpdated", eventDetails );
                     }
     
                     if ( updateEvents ) {
@@ -12905,7 +12808,7 @@ function calendarJs( elementOrId, options, searchOptions ) {
                     removed = true;
     
                     if ( triggerEvent ) {
-                        triggerOptionsEventWithData( "onEventRemoved", event );
+                        fireCustomTrigger( "onEventRemoved", event );
                     }
     
                     if ( updateEvents ) {
@@ -12944,7 +12847,7 @@ function calendarJs( elementOrId, options, searchOptions ) {
             _events = {};
     
             if ( triggerEvent ) {
-                triggerOptionsEvent( "onEventsCleared" );
+                fireCustomTrigger( "onEventsCleared" );
             }
     
             if ( updateEvents ) {
@@ -13137,7 +13040,7 @@ function calendarJs( elementOrId, options, searchOptions ) {
             refreshViews( true, false );
 
             if ( triggerEvent ) {
-                triggerOptionsEvent( "onVisibleEventTypesChanged", _options_Configuration.visibleEventTypes );
+                fireCustomTrigger( "onVisibleEventTypesChanged", _options_Configuration.visibleEventTypes );
             }
         }
 
@@ -13187,7 +13090,7 @@ function calendarJs( elementOrId, options, searchOptions ) {
             } );
 
             if ( triggerEvent ) {
-                triggerOptionsEvent( "onGroupsCleared" );
+                fireCustomTrigger( "onGroupsCleared" );
             }
 
             if ( updateEvents ) {
@@ -13228,7 +13131,7 @@ function calendarJs( elementOrId, options, searchOptions ) {
             } );
 
             if ( triggerEvent ) {
-                triggerOptionsEvent( "onGroupRemoved", groupName );
+                fireCustomTrigger( "onGroupRemoved", groupName );
             }
 
             if ( updateEvents ) {
@@ -13273,7 +13176,7 @@ function calendarJs( elementOrId, options, searchOptions ) {
             refreshViews( true, false );
 
             if ( triggerEvent ) {
-                triggerOptionsEvent( "onVisibleGroupsChanged", _options_Configuration.visibleGroups );
+                fireCustomTrigger( "onVisibleGroupsChanged", _options_Configuration.visibleGroups );
             }
         }
 
@@ -13384,7 +13287,7 @@ function calendarJs( elementOrId, options, searchOptions ) {
      * @returns     {string}                                                The version number.
      */
     this.getVersion = function() {
-        return "2.9.1";
+        return "2.9.2";
     };
 
     /**
@@ -13449,7 +13352,7 @@ function calendarJs( elementOrId, options, searchOptions ) {
             triggerEvent = !isDefinedBoolean( triggerEvent ) ? true : triggerEvent;
 
             if ( triggerEvent ) {
-                triggerOptionsEventWithData( "onOptionsUpdated", _options );
+                fireCustomTrigger( "onOptionsUpdated", _options );
             }
 
             _initialized = false;
@@ -13489,7 +13392,7 @@ function calendarJs( elementOrId, options, searchOptions ) {
             }
     
             if ( triggerEvent ) {
-                triggerOptionsEventWithData( "onSearchOptionsUpdated", _options_Search );
+                fireCustomTrigger( "onSearchOptionsUpdated", _options_Search );
             }
         }
 
@@ -13518,7 +13421,7 @@ function calendarJs( elementOrId, options, searchOptions ) {
             _options.holidays = _options.holidays.concat( holidays );
     
             if ( triggerEvent ) {
-                triggerOptionsEventWithData( "onOptionsUpdated", _options );
+                fireCustomTrigger( "onOptionsUpdated", _options );
             }
     
             if ( updateEvents ) {
@@ -13563,7 +13466,7 @@ function calendarJs( elementOrId, options, searchOptions ) {
             _options.holidays = holidaysRemaining;
 
             if ( triggerEvent ) {
-                triggerOptionsEventWithData( "onOptionsUpdated", _options );
+                fireCustomTrigger( "onOptionsUpdated", _options );
             }
 
             if ( updateEvents ) {